Adding prom-keycloak-proxy metrics with same token
This adds a /prom-keycloak-proxy API that passes the user's same access
token to the actual prom-keycloak-proxy. This allows us to create
frontend dashboards with Plotly JS that access metrics based on
fine-grain resource permissions. Users can access or be denied access to
dashboards and metrics based on their groups in Keycloak.